Add JWT

+++ WebAPIproject/Controllers/ResortController.cs
@@ -1,7 +1,10 @@
 ﻿using DAL.DTO;
 using DAL.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MODELS;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace WebAPIproject.Controllers
 {
@@ -10,11 +13,15 @@ namespace WebAPIproject.Controllers
     public class ResortController : ControllerBase
     {
         private readonly IResortInterface _resortData;
-        public ResortController(IResortInterface resortData)
+        private readonly IAuthInterface _IAuthInterface;
+        private readonly IUserInterface _dbUser;
+        public ResortController(IResortInterface resortData, IAuthInterface IAuthInterface, IUserInterface dbUser)
         {
             _resortData = resortData;
+            _IAuthInterface = IAuthInterface;
+            _dbUser = dbUser;
         }
-        // יצירת ריזורט חדש
+        [Authorize]
         [HttpPost]
         public async Task<IActionResult> CreateResort([FromBody] ResortDTO resort)
         {
@@ -31,16 +38,14 @@ namespace WebAPIproject.Controllers
             }
             return StatusCode(StatusCodes.Status500InternalServerError, "Error creating resort");
         }
-
-        // שליפת כל הריזורטים
+        [Authorize]
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Resort>>> GetAllResorts()
         {
             var resorts = await _resortData.GetAllResorts();
             return Ok(resorts);
         }
-
-        // שליפת ריזורט לפי מזהה (שם הריזורט)
+        [Authorize]
         [HttpGet("{resortName}")]
         public async Task<ActionResult<Resort>> GetResortByName(string resortName)
         {
@@ -52,10 +57,29 @@ namespace WebAPIproject.Controllers
             return Ok(resort);
         }
 
-        // עדכון ריזורט קיים
+        [Authorize]
         [HttpPut("{resortName}")]
         public async Task<ActionResult> UpdateResort(string resortName, [FromBody] ResortDTO updatedResort)
         {
+            var token = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+            var (emailClaim, passwordClaim) = _IAuthInterface.DecodeJwtToken(token);
+
+            if (emailClaim == null || passwordClaim == null)
+            {
+                return Unauthorized();
+            }
+            var user = await _dbUser.GetUserByPasswordAndEmail(passwordClaim, emailClaim);
+
+            if (user == null)
+            {
+                return Unauthorized();
+            }
+            var res = await _resortData.GetResortByName(resortName);
+            // Compare the Phone
+            if (user.phone != res.ownerPhone)
+            {
+                return Forbid();
+            }
             bool result = await _resortData.UpdateResort(resortName, updatedResort);
             if (result)
             {
@@ -64,10 +88,30 @@ namespace WebAPIproject.Controllers
             return BadRequest();
         }
 
-        // מחיקת ריזורט לפי מזהה (שם הריזורט)
+        [Authorize]
         [HttpDelete("{resortName}")]
         public async Task<ActionResult> DeleteResort(string resortName)
         {
+            var token = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+            var (emailClaim, passwordClaim) = _IAuthInterface.DecodeJwtToken(token);
+
+            if (emailClaim == null || passwordClaim == null)
+            {
+                return Unauthorized();
+            }
+            var user = await _dbUser.GetUserByPasswordAndEmail(passwordClaim, emailClaim);
+
+            if (user == null)
+            {
+                return Unauthorized();
+            }
+            var res = await _resortData.GetResortByName(resortName);
+            // Compare the Phone
+            if (user.phone != res.ownerPhone)
+            {
+                return Forbid();
+            }
+
             await _resortData.DeleteResort(resortName);
             return NoContent();
         }
